Radjeno na dodavanju reversa i stavki u revers

+++ Common/Domain/StavkaReversa.cs
@@ -7,7 +7,7 @@ namespace Domen
     {
         public long Rb { get; set; }
         public Revers Revers { get; set; }
-        public int Kolicina { get; set; }
+        public decimal Kolicina { get; set; }
         public Roba Roba { get; set; }
         public decimal IznosStavke { get; set; }
